Enhancing caching, local operations, client interface
- Fileserver: added logic to clean up directories upon exit
- Directory server: now recording the version of the file, including
when creating remote copy. Version is an integer, with initial value 0
the first time the file is created remotely. Also added another
endpoint, UpdateFileVersion, which can be used by the client
(POTENTIALLY)  to inform the directory server of a successful update to
the remote copy. TODO implement this on the client side.
- ClientCache: added recording and checking of versions, corrected 'add'
method to check for file existence without throwing an exception
(incorrect accessing of entry), changed max size to 10. TODO implement a
LRU strategy.
- ClientAPI: added ability to create a new empty text file locally.
Added versioning and some caching implementation in the read and write
methods, though this needs to be tested thoroughly. Added error handling
to the open_file method (so that a file that doesn't exist isn't
attempted to be opened). Removed the close_file method because it seems
to be irrelevant given my implementation.
- Client: added option to create a local, empty text file. Tested and
ensured cleaning up works, including for the client cache. Also added
the refresh_from_filesystem function call, though this should probably
be scheduled as a continuous task (threading may be required).