Verified caching works correctly, versioning updating correctly
There was an issue where the version wasn't being updated correctly
after 1) creating a remote copy and 2) then doing a remote write. This
was because the FILES_ON_RECORD_BY_NAME[] array wasn't being updated
with the fileversion, and this was the array being checked every time by
the client to see whether they could update from the cache or not.