Making creation of new file more robust, and README
Previously, if you wanted to create a file that already existed, the
contents of the existing file would be overwritten and replaced with an
empty file. Now, if the file is there and has contents, it will not be
overwritten. If it isn't there, it will be created. If the directory
isn't there, the directory and the file will be created  (after a prompt
from the user is entered).